# server/docs/swagger.yaml
...
host: localhost:3002/api
info:
  contact: {}
  description: This is the API for Timeful.
  title: Timeful API
  version: "1.0"
paths:
...
  /auth/visitor-identities:
    post:
      consumes:
      - application/json
      parameters:
      - description: Browser-local public identities; matching HttpOnly credentials
          are required
        in: body
        name: payload
        required: true
        schema:
          properties:
            identities:
              items:
                properties:
                  eventId:
                    type: string
                  eventVisitorId:
                    type: string
                type: object
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
      summary: Associate browser Event Visitor Identities with the authenticated account
      tags:
      - auth
...
swagger: "2.0"